backend/pyproject.toml
[tool.poetry]
name = "smart-gallery-backend"
version = "0.1.0"
description = "Backend-сервер для управления фотогалереей Smart Gallery, разработанный с использованием FastAPI. Обеспечивает загрузку изображений, поиск по эмбеддингам и управление галереей через API."
authors = ["GrishaTS <bezgrisa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
aiosqlite = "^0.21.0"
sqlalchemy = "^2.0.38"
greenlet = "^3.1.1"
pydantic-settings = "^2.7.1"
python-multipart = "^0.0.20"
pillow = "^11.1.0"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
